services/notebook-processor/requirements.txt
aiofiles~=24.1.0
ipython~=8.26.0
ipykernel~=6.29.5
jinja2~=3.1.4
jupytext==1.16.3
nats-py~=2.8.0
nbformat~=5.10.4
nbconvert~=7.16.4
matplotlib
numpy
pandas
torch
torchvision
torchaudio
tqdm
scikit-learn
sqlalchemy
numba
cython
seaborn
toolz
skorch
scipy
fastai
pydantic